Remove alarm boot check when alarm time is after current time
If there are two alarms, one is snoozed alarm and the other is going to
fire. When boot-complete, these two alarms will be registered again.
One of them will be set as power off alarm to fire. And the other one will
be kept as original state. And once the firing one is snoozed or dismissed,
the other one will be set as next power off alarm. But its time is missed,
the power off alarm will be disabled.

And it is not necessary to add alarm boot check here because power off alarm
instance will be set to fired state after setting to missed state in current
design.

Change-Id: I48527bdb7ca1ffbe4d95e3234e948d5e3d1e4940
CRs-Fixed: 1063863